Update testapplication.cpp
updated the test application that it also demonstrates the biweight midvariance estimator. 
And an example was added where an even larger number of outliers have been added. 
With the larger point number, one can see that it computes the estimators (which all should break down because the number of outliers in this example is larger than the number of in-liers) with the ransac.