Add files via upload
update: the source files now come with a CMakelists.txt file. The user must insert the paths to OpenCV and cfitsio include and lib directories. 
Then CMake should handle the rest of the platform differences, at least in theory. Compiling and linking with CMake was tested on windows. I still have to test it on Linux and Mac.